+++ Cargo.toml
@@ -1,7 +1,7 @@
 [workspace]
 members = [
-    "hyperspectra-cli",
-    "hyperspectra"
+    "vanadium-cli",
+    "vanadium-core"
 ]
 
 [profile.release]
