+++ benchmark.sh
@@ -4,14 +4,14 @@ DEVICE='workstation'
 IN_DIR='/data/bench-data'
 OUT_DIR='.'
 
-PCA_SMALL="hyperspectra-cli pca --csv -i $IN_DIR/small-{type} -h $IN_DIR/small-{type}.hdr -o $OUT_DIR/pca.{type}.csv -d 3 --min 0.0 --max 1.0"
-PCA_MED="hyperspectra-cli pca --csv -i $IN_DIR/medium-{type} -h $IN_DIR/medium-{type}.hdr -o $OUT_DIR/pca.{type}.csv -d 3 --min 0.0 --max 1.0"
+PCA_SMALL="vanadium-cli pca --csv -i $IN_DIR/small-{type} -h $IN_DIR/small-{type}.hdr -o $OUT_DIR/pca.{type}.csv -d 3 --min 0.0 --max 1.0"
+PCA_MED="vanadium-cli pca --csv -i $IN_DIR/medium-{type} -h $IN_DIR/medium-{type}.hdr -o $OUT_DIR/pca.{type}.csv -d 3 --min 0.0 --max 1.0"
 
 WARMUP="rm $OUT_DIR/*.png $OUT_DIR/*bil $OUT_DIR/*bip $OUT_DIR/*bsq $OUT_DIR/*.csv || true"
 COLD_UP="rm $OUT_DIR/*.png $OUT_DIR/*bil $OUT_DIR/*bip $OUT_DIR/*bsq $OUT_DIR/*.csv || true; sync; echo 3 | tee /proc/sys/vm/drop_caches"
 
-CONVERT_IN_SMALL="hyperspectra-cli convert -i $IN_DIR/small-{type} -n $IN_DIR/small-{type}.hdr"
-CONVERT_IN_MED="hyperspectra-cli convert -i $IN_DIR/medium-{type} -n $IN_DIR/medium-{type}.hdr"
+CONVERT_IN_SMALL="vanadium-cli convert -i $IN_DIR/small-{type} -n $IN_DIR/small-{type}.hdr"
+CONVERT_IN_MED="vanadium-cli convert -i $IN_DIR/medium-{type} -n $IN_DIR/medium-{type}.hdr"
 
 CONVERT_OUT_BIP="-o $OUT_DIR/out-bip -t bip"
 CONVERT_OUT_BIL="-o $OUT_DIR/out-bil -t bil"
