+++ vanadium-cli/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "hyperspectra-cli"
+name = "vanadium-cli"
 version = "1.0.0-alpha"
 authors = ["noah <nomaxx117@gmail.com>"]
 edition = "2018"
@@ -17,8 +17,8 @@ jemallocator = "0.3.2"
 version = "0.3.1"
 default-features = false
 
-[dependencies.hyperspectra]
-path = "../hyperspectra"
+[dependencies.vanadium-core]
+path = "../vanadium-core"
 
 [dependencies.structopt]
 version = "0.3.21"
