+++ vanadium-cli/src/convert.rs
@@ -2,9 +2,9 @@ use std::error::Error;
 use std::fs::{File, OpenOptions, read_to_string};
 use std::str::FromStr;
 
-use hyperspectra::container::{convert, IterableImage, IterableImageMut, LockImage};
-use hyperspectra::container::mapped::{Bip, Bsq};
-use hyperspectra::header::{Headers, Interleave};
+use vanadium_core::container::{convert, IterableImage, IterableImageMut, LockImage};
+use vanadium_core::container::mapped::{Bip, Bsq};
+use vanadium_core::header::{Headers, Interleave};
 
 use crate::cli::ConvertOpt;
