+++ vanadium-cli/src/main.rs
