+++ vanadium-core/src/bar.rs
