+++ vanadium-core/src/container/mapped/bip/bip_iter.rs
