+++ vanadium-core/src/container/mapped/bip/bip_iter_mut.rs
