+++ vanadium-core/src/container/mapped/bip/mod.rs
