+++ vanadium-core/src/container/render.rs
