+++ vanadium-core/src/container/stat/mod.rs
