+++ vanadium-core/src/container/stat/tests.rs
