+++ vanadium-core/src/header.rs
